refactor : Paths 클래스를 constants 패키지로 이동
상수 클래스들이 들어있는 패키지 생성
