+++ src/main/java/webserver/httpUtils/RequestGetter.java
@@ -2,7 +2,7 @@ package webserver.httpUtils;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.Paths;
+import webserver.constants.Paths;
 import webserver.httpUtils.entity.ReqLine;
 
 import java.io.*;
