+++ src/main/java/webserver/service/LogInService.java
@@ -8,7 +8,7 @@ import db.UserIdSession;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.Paths;
+import webserver.constants.Paths;
 import webserver.httpUtils.Request;
 import webserver.httpUtils.Response;
