+++ src/main/java/webserver/service/SignUpService.java
@@ -5,12 +5,10 @@ import db.Database;
 import model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import webserver.Paths;
+import webserver.constants.Paths;
 import webserver.httpUtils.Request;
 import webserver.httpUtils.Response;
 
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.Map;
 
 public class SignUpService implements Service{
